+++ lib/puppet/type/wls_deployment.rb
@@ -49,8 +49,8 @@ module Puppet
     property :targettype
     property :deploymenttype
     property :versionidentifier
-    property :remote
-    property :upload
+    parameter :remote
+    parameter :upload
     property :stagingmode
 
     add_title_attributes(:deployment_name) do